# light_render: Clean up and optimize RenderTriangle